renaming edition files to edit
